Allow existing timeline events files to be destroyed using 'mark for deletion'.

+++ app/assets/javascripts/components/file.js.jsx
@@ -1,25 +1,57 @@
 var File = React.createClass({
   propTypes: {
-    name: React.PropTypes.string,
-    identifier: React.PropTypes.string,
-    private: React.PropTypes.bool
+    name: React.PropTypes.string.isRequired,
+    identifier: React.PropTypes.string.isRequired,
+    private: React.PropTypes.bool.isRequired,
+    persisted: React.PropTypes.bool,
+    deleteFileCB: React.PropTypes.func,
+    markFileForDeletionCB: React.PropTypes.func,
+    markedForDeletion: React.PropTypes.bool
   },
 
   deleteFile: function () {
-    // Remove the corresponding input element.
-    $('#timeline_event_file_' + this.props.identifier).remove();
+    if (confirm('Are you sure you want to delete this file?')) {
+      // Remove the corresponding input element.
+      $('#timeline_event_file_' + this.props.identifier).remove();
 
-    // Handle passed all the way from attachments editor.
-    this.props.deleteFileCB(this.props.index);
+      // Handle passed all the way from attachments editor.
+      this.props.deleteFileCB(this.props.index);
+    }
+  },
+
+  markFileForDeletion: function () {
+    if (confirm('This file will be marked for deletion, and removed when you submit changes to timeline event. Continue?')) {
+      this.props.markFileForDeletionCB(this.props.index);
+    }
+  },
+
+  showDeleteButton: function () {
+    return !this.props.markedForDeletion;
+  },
+
+  fileIconClasses: function() {
+    if (this.props.markedForDeletion) {
+      return 'fa fa-trash';
+    } else if (this.props.private) {
+      return 'fa fa-user-secret';
+    } else {
+      return 'fa fa-file-o';
+    }
   },
 
   render: function () {
     return (
       <li className="list-group-item">
-        <i className={ this.props.private ? 'fa fa-user-secret' : 'fa fa-file-o'}/>
-        &nbsp;{ this.props.name }
+        <i className={ this.fileIconClasses() }/>
+        &nbsp;
+        <span className={this.props.markedForDeletion ? 'strike' : ''}>{ this.props.name }</span>
         <div className="pull-right">
-          <a className="margin-right-10" onClick={this.deleteFile}>Delete</a>
+          { this.showDeleteButton() &&
+          <a className="margin-right-10" onClick={this.props.persisted ? this.markFileForDeletion : this.deleteFile}>Delete</a>
+          }
+          { this.props.markedForDeletion &&
+          <em>Marked for Deletion</em>
+          }
         </div>
       </li>
     );
